@@ -1,4 +1,4 @@
-class validationUtils{
+class ValidationUtils{
   static bool isValidEmail(String? email){
     if(email ==null) return false;
     var regex = RegExp(r"^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,253}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,253}[a-zA-Z0-9])?)*$");
