@@ -1,8 +1,5 @@
-import 'package:ecommerce/domain/model/AuthResultDto.dart';
-
-import '../../User.dart';
-
-
+import 'package:ecommerce/data/model/UserDTO.dart';
+import 'package:ecommerce/domain/model/AuthResult.dart';
 /// message : "success"
 /// user : {"name":"bashar walid","email":"basharwalid2002@gmail.com","role":"user"}
 /// token : "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY0YTM4NmY5OTVmY2UwNDQxZDc4Y2ZmMSIsIm5hbWUiOiJiYXNoYXIgd2FsaWQiLCJyb2xlIjoidXNlciIsImlhdCI6MTY4ODQ0MDcwNiwiZXhwIjoxNjk2MjE2NzA2fQ.jj4lQFLCZjYfkXd1Ir5qJDEDRN-dGnndvRqG4EKsMeQ"
@@ -15,12 +12,12 @@ class LoginResponse {
 
   LoginResponse.fromJson(dynamic json) {
     message = json['message'];
-    user = json['user'] != null ? User.fromJson(json['user']) : null;
+    user = json['user'] != null ? UserDTO.fromJson(json['user']) : null;
     token = json['token'];
     statusMsg = json['statusMsg'];
   }
   String? message;
-  User? user;
+  UserDTO? user;
   String? token;
   String? statusMsg;
 
@@ -37,8 +34,8 @@ class LoginResponse {
     return statusMsg != 'fail';
   }
 
-  AuthResultDto toAuthResultDto(){
-  return AuthResultDto(
+  AuthResult toDomain(){
+  return AuthResult(
     user: user?.toUserDTo(),
     token: token,
   );
