@@ -1,4 +1,4 @@
-import 'package:ecommerce/data/dataSource/categories_online_dataSource_impl.dart';
+
 import 'package:ecommerce/domain/model/CategoryResultDto.dart';
 import 'package:ecommerce/domain/repository/Categories_repository.dart';
 
@@ -8,7 +8,7 @@ class CategoriesRepositoryimpl implements CategoriesRepository{
   CategoriesOnlineDataSource dataSource;
   CategoriesRepositoryimpl(this.dataSource);
   @override
-  Future<CategoryResultDto> getCategories({int page = 1, int limit = 20, String? categorySlug}) {
+  Future<CategoryResult> getCategories({int page = 1, int limit = 20, String? categorySlug}) {
     return dataSource.getCategories(
       page: page,
       limit: limit,
