@@ -1,9 +1,5 @@
-/// currentPage : 1
-/// numberOfPages : 1
-/// limit : 40
-
-class PaginationDto {
-  PaginationDto({
+class Pagination {
+  Pagination({
       this.currentPage, 
       this.numberOfPages, 
       this.limit,});
