@@ -1,5 +1,5 @@
 import 'package:ecommerce/domain/model/CategoryResultDto.dart';
 
 abstract class CategoriesRepository{
-  Future<CategoryResultDto> getCategories({int page =1 , int limit = 20 , String? categorySlug });
+  Future<CategoryResult> getCategories({int page =1 , int limit = 20 , String? categorySlug });
 }
