@@ -1,22 +1,22 @@
+import 'package:ecommerce/Core/DI/di.dart';
+import 'package:ecommerce/Core/providers/Auth_provider.dart';
+import 'package:ecommerce/Core/utils/validation%20utils.dart';
+import 'package:ecommerce/domain/usecase/register_usecase.dart';
 import 'package:ecommerce/my_theme.dart';
-import 'package:ecommerce/providers/Auth_provider.dart';
 import 'package:ecommerce/ui/home/home_screen.dart';
 import 'package:ecommerce/ui/login/login_screen.dart';
 import 'package:ecommerce/ui/register/register_view_model.dart';
 import 'package:flutter/material.dart';
-import 'package:ecommerce/utils/validation utils.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import '../components/Custom_form_field.dart';
 
+// ignore: must_be_immutable
 class RegisterScreen extends StatelessWidget {
   static const String routeName = 'register screen';
-  var formkey = GlobalKey<FormState>();
-  var namecontroller = TextEditingController(text: "bashar walid");
-  var Emailcontroller = TextEditingController(text: "basharwalid2002@gmail.com");
-  var mobilecontroller = TextEditingController(text: "01016309078");
-  var Passwordcontroller = TextEditingController(text: "123456");
-  var PasswordConfirmcontroller = TextEditingController(text: "123456");
-  var viewModel = RegisterViewModel();
+
+  var viewModel = RegisterViewModel(RegisterUseCase(injectAuthRepo()));
+
+  RegisterScreen({super.key});
   @override
   Widget build(BuildContext context) {
     return BlocConsumer<RegisterViewModel , RegisterViewState>
@@ -52,13 +52,13 @@ class RegisterScreen extends StatelessWidget {
                   const SizedBox(height: 80,),
                   Image.asset("assets/images/route_logo_big.png"),
                   Form(
-                    key: formkey,
+                    key: viewModel.formkey,
                     child: Column(
                       crossAxisAlignment: CrossAxisAlignment.stretch,
                       children: [
                         CustomFormField(
                           hint: "Enter your Full name",
-                          controller: namecontroller,
+                          controller: viewModel.namecontroller,
                           label: "Full name",
                           validator:(text){
                             if(text==null || text.trim().isEmpty){
@@ -69,14 +69,14 @@ class RegisterScreen extends StatelessWidget {
                         ),
                         CustomFormField(
                           hint: "enter your email address",
-                          controller: Emailcontroller,
+                          controller: viewModel.Emailcontroller,
                           label: "Email",
                           keyboardtype: TextInputType.emailAddress,
                           validator: (text) {
                             if(text==null || text.trim().isEmpty){
                               return 'please Enter email';
                             }
-                            if (!validationUtils.isValidEmail(text))
+                            if (!ValidationUtils.isValidEmail(text))
                             {
                               return "please enter a valid email";
                             }
@@ -85,7 +85,7 @@ class RegisterScreen extends StatelessWidget {
                         ),
                         CustomFormField(
                           hint: "enter your Mobile number",
-                          controller: mobilecontroller,
+                          controller: viewModel.mobilecontroller,
                           label: "Mobile no.",
                           keyboardtype: TextInputType.phone,
                           validator: (text) {
@@ -106,7 +106,7 @@ class RegisterScreen extends StatelessWidget {
                             }
                             return null;
                           },
-                          controller: Passwordcontroller,
+                          controller:viewModel.Passwordcontroller,
                           label: "Password",
                           ispassword: true,
                         ),
@@ -116,12 +116,12 @@ class RegisterScreen extends StatelessWidget {
                             if(text==null || text.trim().isEmpty){
                               return 'please Enter Password confirmation';
                             }
-                            if(Passwordcontroller.text != text){
+                            if(viewModel.Passwordcontroller.text != text){
                               return "Password doesn't match";
                             }
                             return null;
                           },
-                          controller: PasswordConfirmcontroller,
+                          controller: viewModel.PasswordConfirmcontroller,
                           label: "Confirm Password",
                           ispassword: true,
                         ),
@@ -170,9 +170,9 @@ class RegisterScreen extends StatelessWidget {
       ;
   }
   void register() async{
-    if(formkey.currentState!.validate()== false){
+    if(viewModel.formkey.currentState!.validate()== false){
       return;
     }
-    viewModel.Register(namecontroller.text, Emailcontroller.text, Passwordcontroller.text, PasswordConfirmcontroller.text, mobilecontroller.text);
+    viewModel.Register(viewModel.namecontroller.text, viewModel.Emailcontroller.text, viewModel.Passwordcontroller.text, viewModel.PasswordConfirmcontroller.text, viewModel.mobilecontroller.text);
   }
 }
